update structure to remove init params
