+++ resources_measurement/linux_resources/cgroup_versions/common_paths.py
@@ -1,0 +1,3 @@
+# A cgroup is a feature that allows you to allocate, limit, and monitor system resources among user-defined groups of processes.
+# Enables control over resource distribution, ensuring that no single group can monopolize the system resources.
+SYSTEM_CGROUP_DIR_PATH = r"/sys/fs/cgroup/"
